Restore invariant after creating a revision
This fixes a scenario like this:

o Consider size-3-clause XYZ.
o The assignment contains -Y and -Z, hence X is propagated.
o Then some other reason clause is deleted. The cone contains -Z and X.
o As X is in the cone, we need to revise the watches in XYZ.

Previously this would result in the configuration ZYX where
watch Z is unassigned and watch Y is falsified, thus Invariant 1 is violated.
We want XZY instead.